% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ee_col}
\alias{ee_col}
\title{Create a colored tile plot using ggplot2.}
\usage{
ee_col(data, x, y, z, xlab = x, ylab = y)
}
\arguments{
\item{data}{A data frame.}

\item{x}{The variable to be plotted on the x-axis.}

\item{y}{The variable to be plotted on the y-axis.}

\item{z}{The variable for color fill.}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis.}
}
\value{
A ggplot2 colored tile plot.
}
\description{
This function creates a colored tile plot using ggplot2 based on the specified data
and aesthetics.
}
