% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ee_density}
\alias{ee_density}
\title{Create a density plot using ggplot2.}
\usage{
ee_density(data, x, xlab = x, ylab = "Density", fill = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{x}{The variable to be plotted on the x-axis.}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis (default is "Density").}

\item{fill}{A variable for fill grouping (optional).}
}
\value{
A ggplot2 density plot.
}
\description{
This function creates a density plot using ggplot2 based on the specified data
and aesthetics.
}
