% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ee_freqpoly}
\alias{ee_freqpoly}
\title{Create a frequency polygon plot using ggplot2.}
\usage{
ee_freqpoly(data, x, xlab = x, ylab = "Frequency", color = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{x}{The variable to be plotted on the x-axis.}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis (default is "Frequency").}

\item{color}{A variable for color grouping (optional).}
}
\value{
A ggplot2 frequency polygon plot.
}
\description{
This function creates a frequency polygon plot using ggplot2 based on the specified data
and aesthetics.
}
