% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ee_hex}
\alias{ee_hex}
\title{Create a hexbin plot using ggplot2.}
\usage{
ee_hex(data, x, y, xlab = x, ylab = y)
}
\arguments{
\item{data}{A data frame.}

\item{x}{The variable to be plotted on the x-axis.}

\item{y}{The variable to be plotted on the y-axis.}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis.}
}
\value{
A ggplot2 hexbin plot.
}
\description{
This function creates a hexbin plot using ggplot2 based on the specified data
and aesthetics.
}
