% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ee_line}
\alias{ee_line}
\title{Create a line plot using ggplot2.}
\usage{
ee_line(data, x, y, xlab = x, ylab = y, color = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{x}{The variable to be plotted on the x-axis.}

\item{y}{The variable to be plotted on the y-axis.}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis.}

\item{color}{A variable for color grouping (optional).}
}
\value{
A ggplot2 line plot.
}
\description{
This function creates a line plot using ggplot2 based on the specified data
and aesthetics.
}
