% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ee_rug}
\alias{ee_rug}
\title{Create a rug plot using ggplot2.}
\usage{
ee_rug(data, x, xlab = x, ylab = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{x}{The variable to be plotted on the x-axis.}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis (optional).}
}
\value{
A ggplot2 rug plot.
}
\description{
This function creates a rug plot using ggplot2 based on the specified data
and aesthetics.
}
