% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ee_violin}
\alias{ee_violin}
\title{Create a violin plot using ggplot2.}
\usage{
ee_violin(data, x, y, xlab = x, ylab = y, fill = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{x}{The variable to be plotted on the x-axis.}

\item{y}{The variable to be plotted on the y-axis.}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis.}

\item{fill}{A variable for fill grouping (optional).}
}
\value{
A ggplot2 violin plot.
}
\description{
This function creates a violin plot using ggplot2 based on the specified data
and aesthetics.
}
