% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{get_labels}
\alias{get_labels}
\title{Helper function for plotting.}
\usage{
get_labels(xlab, ylab)
}
\arguments{
\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis.}
}
\value{
A list of ggplot2 objects.
}
\description{
Helper function for plotting.
}
